:recycle: Refactor: main 내 Component 테스트 위젯 ComponentTestScreen으로 분리
